Cleanup, and directories / documentation
Added || die lines for proper handling and debuging if/when there's a
problem.  Also filled out src_install to create $HOME/cricket-1.0.8 and
symlink $HOME/cricket -> cricket-1.0.8.  Also installing all of the html
docs in doc.  Currently it's installing to /, I need to figure out how
to move it to /home/cricket/cricket-1.0.8